Move CMyTabCtrl to individual files
