+++ tabs_too/About_Tab_Controls/About_Tab_Controls.vcxproj
@@ -185,6 +185,7 @@
   <ItemGroup>
     <ClInclude Include="About_Tab_Controls.h" />
     <ClInclude Include="About_Tab_ControlsDlg.h" />
+    <ClInclude Include="CMyTabCtrl.h" />
     <ClInclude Include="framework.h" />
     <ClInclude Include="pch.h" />
     <ClInclude Include="Resource.h" />
@@ -194,6 +195,7 @@
   <ItemGroup>
     <ClCompile Include="About_Tab_ControlsApp.cpp" />
     <ClCompile Include="About_Tab_ControlsDlg.cpp" />
+    <ClCompile Include="CMyTabCtrl.cpp" />
     <ClCompile Include="pch.cpp">
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Debug|x64'">Create</PrecompiledHeader>
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Debug|Win32'">Create</PrecompiledHeader>
