+++ tabs_too/About_Tab_Controls/About_Tab_Controls.vcxproj.filters
@@ -36,6 +36,9 @@
     <ClInclude Include="TabPage.h">
       <Filter>Header Files</Filter>
     </ClInclude>
+    <ClInclude Include="CMyTabCtrl.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="About_Tab_ControlsApp.cpp">
@@ -48,6 +51,9 @@
       <Filter>Source Files</Filter>
     </ClCompile>
     <ClCompile Include="TabPage.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="CMyTabCtrl.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
   </ItemGroup>
