+++ tabs_too/About_Tab_Controls/pch.h
@@ -10,6 +10,7 @@
 // add headers that you want to pre-compile here
 #include "framework.h"
 
+#include <iostream>
 #include "DebuggingConsole.h"
 
 #endif //PCH_H
